backend/requirements.txt
torch
torchvision
timm
numpy
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
requests
facenet-pytorch
scikit-image
Pillow
pgvector
cmake
dlib
gunicorn
geoalchemy2

